+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Move USB remapping for STM32F042x to Cargo.toml as `usb-remap` feature.
+
 ## [v0.17.0] - 2020-06-27
 
 ### Changed